This calculates beta values for lineal regression
The thing is that I used two scripts, one for get the strings to use properly the jama library and the other one to calculate the beta values. For y cost I need to use Manhattan distance cost. I will ask you to help me with that function so I can use it, I think the cost is in time and that's why I didn't did it by my self.